WIP: extract shared docker/btrfs/snapshot toolkit into mngr_vps_docker.container_setup
Move the reusable outer-host container, btrfs, snapshot-helper, and image-build
helpers out of VpsDockerProvider into a new container_setup module with public
names, plus shared setup_container_ssh and build_image_on_outer_from_build_args
orchestration functions. instance.py now imports them; vps_docker behavior is
unchanged (all 202 unit tests pass). This is the shared surface mngr_lima will
import for its docker-in-VM mode.
