+++ libs/mngr_vps_docker/imbue/mngr_vps_docker/_snapshot_helper_test.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import pytest
 
-from imbue.mngr_vps_docker.instance import _load_resource_text
+from imbue.mngr_vps_docker.container_setup import load_resource_text
 
 
 @pytest.mark.skipif(shutil.which("bash") is None, reason="bash not on PATH")
@@ -16,11 +16,11 @@ def test_snapshot_helper_script_passes_bash_syntax_check(tmp_path: Path) -> None
     """`bash -n` must accept the bundled snapshot_helper.sh as valid syntax.
 
     Also exercises the wheel's resource bundling end-to-end: if the
-    pyproject.toml `include` directive ever drops the .sh file, `_load_resource_text`
+    pyproject.toml `include` directive ever drops the .sh file, `load_resource_text`
     raises before we even get to the bash check.
     """
     script_path = tmp_path / "snapshot_helper.sh"
-    script_path.write_text(_load_resource_text("snapshot_helper.sh"))
+    script_path.write_text(load_resource_text("snapshot_helper.sh"))
     result = subprocess.run(
         ["bash", "-n", str(script_path)],
         capture_output=True,
@@ -36,4 +36,4 @@ def test_snapshot_helper_unit_is_loadable() -> None:
     Combined with the bash-syntax test above, this ensures both resource
     files survive the wheel build.
     """
-    _load_resource_text("snapshot_helper.service")
+    load_resource_text("snapshot_helper.service")
