+++ libs/mngr_vps_docker/imbue/mngr_vps_docker/instance.py
@@ -1,9 +1,6 @@
 import json
 import os
-import re
 import shlex
-import shutil
-import tempfile
 import time
 from collections.abc import Callable
 from collections.abc import Iterator
@@ -12,10 +9,8 @@ from collections.abc import Sequence
 from contextlib import contextmanager
 from datetime import datetime
 from datetime import timezone
-from importlib import resources as importlib_resources
 from pathlib import Path
 from typing import Any
-from typing import Final
 
 from loguru import logger
 from pydantic import ConfigDict
@@ -61,7 +56,6 @@ from imbue.mngr.primitives import AgentName
 from imbue.mngr.primitives import CommandString
 from imbue.mngr.primitives import DiscoveredAgent
 from imbue.mngr.primitives import DiscoveredHost
-from imbue.mngr.primitives import DockerBuilder
 from imbue.mngr.primitives import HostId
 from imbue.mngr.primitives import HostName
 from imbue.mngr.primitives import HostState
@@ -75,36 +69,49 @@ from imbue.mngr.primitives import VolumeId
 from imbue.mngr.providers.base_provider import BaseProviderInstance
 from imbue.mngr.providers.listing_utils import build_listing_collection_script
 from imbue.mngr.providers.listing_utils import parse_listing_collection_output
-from imbue.mngr.providers.ssh_host_setup import build_add_authorized_keys_command
-from imbue.mngr.providers.ssh_host_setup import build_add_known_hosts_command
-from imbue.mngr.providers.ssh_host_setup import build_check_and_install_packages_command
-from imbue.mngr.providers.ssh_host_setup import build_configure_ssh_command
 from imbue.mngr.providers.ssh_host_setup import build_start_activity_watcher_command
 from imbue.mngr.providers.ssh_utils import add_host_to_known_hosts
 from imbue.mngr.providers.ssh_utils import create_pyinfra_host
 from imbue.mngr.providers.ssh_utils import load_or_create_host_keypair
 from imbue.mngr.providers.ssh_utils import load_or_create_ssh_keypair
 from imbue.mngr.providers.ssh_utils import wait_for_sshd
-from imbue.mngr.utils.git_utils import rsync_worktree_over_clone
 from imbue.mngr_vps_docker.cloud_init import generate_cloud_init_user_data
 from imbue.mngr_vps_docker.config import VpsDockerProviderConfig
-from imbue.mngr_vps_docker.errors import VpsProvisioningError
-from imbue.mngr_vps_docker.host_store import AGENTS_SUBPATH
+from imbue.mngr_vps_docker.container_setup import CONTAINER_ENTRYPOINT_CMD
+from imbue.mngr_vps_docker.container_setup import HOST_DIR_SUBPATH
+from imbue.mngr_vps_docker.container_setup import HOST_VOLUME_MOUNT_PATH
+from imbue.mngr_vps_docker.container_setup import LABEL_HOST_ID
+from imbue.mngr_vps_docker.container_setup import LABEL_HOST_NAME
+from imbue.mngr_vps_docker.container_setup import LABEL_PROVIDER
+from imbue.mngr_vps_docker.container_setup import LABEL_TAGS
+from imbue.mngr_vps_docker.container_setup import SNAPSHOT_READ_MOUNT_PATH
+from imbue.mngr_vps_docker.container_setup import SNAPSHOT_TRIGGER_MOUNT_PATH
+from imbue.mngr_vps_docker.container_setup import build_image_on_outer_from_build_args
+from imbue.mngr_vps_docker.container_setup import check_file_exists_on_outer
+from imbue.mngr_vps_docker.container_setup import commit_container
+from imbue.mngr_vps_docker.container_setup import create_bind_volume_on_outer
+from imbue.mngr_vps_docker.container_setup import delete_btrfs_subvolume_on_outer
+from imbue.mngr_vps_docker.container_setup import docker_inspect_running
+from imbue.mngr_vps_docker.container_setup import exec_in_container
+from imbue.mngr_vps_docker.container_setup import host_volume_name_for
+from imbue.mngr_vps_docker.container_setup import prepare_btrfs_on_outer
+from imbue.mngr_vps_docker.container_setup import provision_snapshot_helper_on_outer
+from imbue.mngr_vps_docker.container_setup import pull_image
+from imbue.mngr_vps_docker.container_setup import remove_container
+from imbue.mngr_vps_docker.container_setup import remove_host_from_known_hosts
+from imbue.mngr_vps_docker.container_setup import remove_volume
+from imbue.mngr_vps_docker.container_setup import run_container
+from imbue.mngr_vps_docker.container_setup import run_docker
+from imbue.mngr_vps_docker.container_setup import seed_host_volume_layout_on_outer
+from imbue.mngr_vps_docker.container_setup import setup_container_ssh
+from imbue.mngr_vps_docker.container_setup import snapshot_trigger_volume_name_for
+from imbue.mngr_vps_docker.container_setup import start_container
+from imbue.mngr_vps_docker.container_setup import stop_container
 from imbue.mngr_vps_docker.host_store import VpsDockerHostRecord
 from imbue.mngr_vps_docker.host_store import VpsHostConfig
 from imbue.mngr_vps_docker.host_store import open_host_store
 from imbue.mngr_vps_docker.primitives import VpsInstanceId
 from imbue.mngr_vps_docker.vps_client import VpsClientInterface
-
-
-def _remove_host_from_known_hosts(known_hosts_path: Path, hostname: str, port: int) -> None:
-    """Remove a host entry from the known_hosts file."""
-    if not known_hosts_path.exists():
-        return
-    host_pattern = hostname if port == 22 else f"[{hostname}]:{port}"
-    lines = known_hosts_path.read_text().splitlines(keepends=True)
-    filtered = [line for line in lines if not line.startswith(f"{host_pattern} ")]
-    known_hosts_path.write_text("".join(filtered))
 
 
 class ParsedVpsBuildOptions(FrozenModel):
@@ -175,101 +182,6 @@ def _parse_build_args(
     )
 
 
-def _resolve_dockerfile_paths(
-    docker_build_args: Sequence[str],
-    remote_build_dir: str,
-) -> tuple[str, ...]:
-    """Rewrite relative --file/--dockerfile paths to absolute paths on the VPS.
-
-    Docker resolves --file relative to the daemon's CWD, not the build context.
-    Since the build context was uploaded to remote_build_dir on the VPS, any
-    relative Dockerfile paths must be prefixed with that directory.
-
-    Handles both ``--file=Dockerfile`` and ``-f Dockerfile`` forms.
-    """
-    resolved: list[str] = []
-    is_next_arg_dockerfile = False
-    for arg in docker_build_args:
-        if is_next_arg_dockerfile:
-            if not arg.startswith("/"):
-                arg = f"{remote_build_dir}/{arg}"
-            is_next_arg_dockerfile = False
-        elif arg in ("-f", "--file", "--dockerfile"):
-            is_next_arg_dockerfile = True
-        else:
-            for prefix in ("--file=", "-f=", "--dockerfile="):
-                if arg.startswith(prefix):
-                    dockerfile_path = arg[len(prefix) :]
-                    if not dockerfile_path.startswith("/"):
-                        arg = f"{prefix}{remote_build_dir}/{dockerfile_path}"
-                    break
-        resolved.append(arg)
-    return tuple(resolved)
-
-
-def _emit_docker_build_output(line: str) -> None:
-    """Log a line of docker build output at BUILD level."""
-    stripped = line.strip()
-    if stripped:
-        logger.log(LogLevel.BUILD.value, "{}", stripped, source="docker")
-
-
-# Label constants (same scheme as Docker provider)
-LABEL_PREFIX: Final[str] = "com.imbue.mngr."
-LABEL_PROVIDER: Final[str] = f"{LABEL_PREFIX}provider"
-LABEL_HOST_ID: Final[str] = f"{LABEL_PREFIX}host-id"
-LABEL_HOST_NAME: Final[str] = f"{LABEL_PREFIX}host-name"
-LABEL_TAGS: Final[str] = f"{LABEL_PREFIX}tags"
-
-# Default image when no user customization
-DEFAULT_IMAGE: Final[str] = "debian:bookworm-slim"
-
-# Path inside the agent container where the unified host volume is mounted.
-# The container sees three top-level entries under this mount: host_state.json,
-# agents/, and host_dir/. The container's mngr host_dir symlink resolves into
-# the host_dir/ subdirectory so all of the agent's writes end up on the volume.
-HOST_VOLUME_MOUNT_PATH: Final[str] = "/mngr-vol"
-
-# Subdirectory inside the unified volume that backs the agent's mngr host_dir.
-HOST_DIR_SUBPATH: Final[str] = "host_dir"
-
-# Shell command for the agent container's PID 1: trap SIGTERM and stay alive
-# until SIGTERM arrives so `docker stop` (idle timeout, manual stop) exits cleanly.
-CONTAINER_ENTRYPOINT_CMD: Final[str] = "trap 'exit 0' TERM; tail -f /dev/null & wait"
-
-# In-container path the host_backup service writes / reads snapshot
-# request and result JSON to. Backed by the per-host docker volume
-# ``mngr-snapshot-trigger-<host_id_hex>`` which is also bind-mounted on the
-# outer at ``/var/lib/mngr-snapshot/`` so the outer-side snapshot helper
-# can watch it.
-SNAPSHOT_TRIGGER_MOUNT_PATH: Final[str] = "/mngr-snapshot"
-
-# In-container path that exposes the outer's <btrfs-mount>/snapshots/
-# directory read-only. host_backup reads ``<this>/current`` after the
-# outer helper produces a snapshot there.
-SNAPSHOT_READ_MOUNT_PATH: Final[str] = "/mngr-snapshots"
-
-# Outer-host paths for the snapshot-helper protocol files. Must match
-# what ``resources/snapshot_helper.sh`` watches.
-OUTER_SNAPSHOT_TRIGGER_DIR: Final[Path] = Path("/var/lib/mngr-snapshot")
-
-# Outer-host install paths for the helper script + systemd unit + env file.
-OUTER_HELPER_SCRIPT_PATH: Final[Path] = Path("/usr/local/sbin/snapshot_helper.sh")
-OUTER_HELPER_SERVICE_PATH: Final[Path] = Path("/etc/systemd/system/snapshot_helper.service")
-OUTER_HELPER_ENV_PATH: Final[Path] = Path("/etc/mngr-snapshot-helper.env")
-OUTER_HELPER_SERVICE_NAME: Final[str] = "snapshot_helper.service"
-
-
-def _host_volume_name_for(host_id: HostId) -> str:
-    """Return the unified Docker volume name for a host."""
-    return f"mngr-host-vol-{host_id.get_uuid().hex}"
-
-
-def _snapshot_trigger_volume_name_for(host_id: HostId) -> str:
-    """Return the per-host snapshot-trigger Docker volume name."""
-    return f"mngr-snapshot-trigger-{host_id.get_uuid().hex}"
-
-
 def _read_host_id_label_from_vps(outer: OuterHostInterface) -> HostId | None:
     """Return the host_id label of the (single) mngr container on this VPS, if any.
 
@@ -304,45 +216,6 @@ def _read_host_id_label_from_vps(outer: OuterHostInterface) -> HostId | None:
     return None
 
 
-# Idempotent install: skip if depot already on PATH, otherwise download to
-# /usr/local/bin via depot.dev's official installer. Run once per build (cheap
-# no-op when already present); avoids needing a separate provisioning step.
-_DEPOT_INSTALL_CMD: Final[str] = "command -v depot >/dev/null 2>&1 || curl -fsSL https://depot.dev/install-cli.sh | sh"
-
-# Env-var assignments whose values are secrets and must be redacted before any
-# remote command string ends up in logs or exception messages.
-_SECRET_ENV_VARS: Final[tuple[str, ...]] = ("DEPOT_TOKEN",)
-_SECRET_ENV_PATTERN: Final[re.Pattern[str]] = re.compile(r"\b(" + "|".join(_SECRET_ENV_VARS) + r")=(?:'[^']*'|\S+)")
-
-# Absolute path on the VPS where rsync stashes partial files between
-# attempts. Lives outside the build context (``/tmp/mngr-build-<id>/``) so
-# partial-transfer state never gets included in the docker build context
-# or copied back to the local repo. Persists across retries so subsequent
-# attempts can resume rather than re-uploading completed bytes.
-_RSYNC_PARTIAL_DIR_REMOTE: Final[str] = "/tmp/mngr-rsync-partial"
-
-# How many times to retry a failed rsync upload before giving up.
-_UPLOAD_MAX_ATTEMPTS: Final[int] = 3
-# Backoff between attempts (entry N is the wait *before* attempt N+1).
-_UPLOAD_RETRY_BACKOFF_SECONDS: Final[tuple[float, ...]] = (5.0, 15.0)
-
-# Substrings in rsync stderr that indicate a transient connection-class
-# failure (broken pipe, dropped TCP, fresh-VPS networking flap). Rsync's
-# own catch-all exit 255 with these messages is what fresh Vultr VPSes
-# produce in the first 30-60s of life. Other rsync errors (permission,
-# protocol mismatch, vanished source) are non-transient and we fail fast.
-_RETRYABLE_RSYNC_PATTERNS: Final[tuple[str, ...]] = (
-    "Broken pipe",
-    "Connection reset by peer",
-    "Connection refused",
-    "Connection timed out",
-    "client_loop",
-    "ssh: connect to host",
-    "kex_exchange_identification",
-    "Network is unreachable",
-)
-
-
 def build_vps_tags(host_id: HostId, provider_name: str, extra_tags_raw: str) -> list[str]:
     """Compose the tag list passed to the VPS create call.
 
@@ -363,690 +236,6 @@ def build_vps_tags(host_id: HostId, provider_name: str, extra_tags_raw: str) -> 
         if stripped:
             tags.append(stripped)
     return tags
-
-
-def _redact_secret_env(remote_command: str) -> str:
-    """Return remote_command with values of known-secret env-var assignments replaced."""
-    return _SECRET_ENV_PATTERN.sub(r"\1=<redacted>", remote_command)
-
-
-def _is_retryable_rsync_error(stderr: str) -> bool:
-    """Return True iff stderr looks like a connection-class rsync failure."""
-    return any(pattern in stderr for pattern in _RETRYABLE_RSYNC_PATTERNS)
-
-
-def _docker_inspect_running(outer: OuterHostInterface, container_name: str) -> bool:
-    """Return True iff a container with the given name is running on outer."""
-    result = outer.execute_idempotent_command(
-        f"docker inspect --format '{{{{.State.Running}}}}' {shlex.quote(container_name)}"
-    )
-    if not result.success:
-        return False
-    return result.stdout.strip().lower() == "true"
-
-
-def _check_file_exists_on_outer(outer: OuterHostInterface, path: Path) -> bool:
-    """Return True iff a file exists on outer."""
-    result = outer.execute_idempotent_command(f"test -f {shlex.quote(str(path))}", timeout_seconds=10.0)
-    return result.success
-
-
-def _check_directory_exists_on_outer(outer: OuterHostInterface, path: Path) -> bool:
-    """Return True iff a directory exists on outer."""
-    result = outer.execute_idempotent_command(f"test -d {shlex.quote(str(path))}", timeout_seconds=10.0)
-    return result.success
-
-
-def _is_btrfs_progs_installed_on_outer(outer: OuterHostInterface) -> bool:
-    """Return True iff ``mkfs.btrfs`` is on the outer's PATH (i.e. btrfs-progs is installed)."""
-    result = outer.execute_idempotent_command(
-        "command -v mkfs.btrfs >/dev/null 2>&1",
-        timeout_seconds=10.0,
-    )
-    return result.success
-
-
-def _install_btrfs_progs_on_outer(outer: OuterHostInterface) -> None:
-    """Install btrfs-progs on the outer via apt-get; raise VpsProvisioningError on failure."""
-    command = (
-        "DEBIAN_FRONTEND=noninteractive apt-get update && "
-        "DEBIAN_FRONTEND=noninteractive apt-get install -y btrfs-progs"
-    )
-    result = outer.execute_idempotent_command(command, timeout_seconds=300.0)
-    if not result.success:
-        raise VpsProvisioningError(f"Failed to install btrfs-progs on outer: stderr={result.stderr.strip()!r}")
-
-
-def _get_outer_free_disk_gb(outer: OuterHostInterface, path: Path) -> int:
-    """Return free space at ``path`` on the outer, floor-divided to whole GiB.
-
-    Reads the available-bytes count with ``df --output=avail -B 1`` and does
-    the GiB conversion (`// (1024 ** 3)`) in Python so the result is always
-    less-than-or-equal-to the true free space. ``df``'s own ``-B 1G`` form
-    rounds up, which would let the caller compute a loop-file size up to
-    ~1 GiB larger than actually available; floor-dividing here keeps the
-    caller's allocation math conservative.
-
-    Runs the ``df | tail`` pipeline under ``bash -c 'set -o pipefail; ...'``
-    so a failing ``df`` (e.g. invalid path, permission issue) is surfaced
-    through the "failed to read" branch rather than being masked by
-    ``tail``'s zero exit code and falling through to the "could not parse"
-    branch.
-    """
-    pipeline = f"set -o pipefail; df --output=avail -B 1 {shlex.quote(str(path))} | tail -n 1"
-    result = outer.execute_idempotent_command(
-        f"bash -c {shlex.quote(pipeline)}",
-        timeout_seconds=10.0,
-    )
-    if not result.success:
-        raise VpsProvisioningError(
-            f"Failed to read free disk space at {path} on outer: stderr={result.stderr.strip()!r}"
-        )
-    raw = result.stdout.strip()
-    try:
-        free_bytes = int(raw)
-    except ValueError as e:
-        raise VpsProvisioningError(f"Could not parse free-space output {raw!r} from df at {path} on outer") from e
-    return free_bytes // (1024**3)
-
-
-def _is_path_mounted_on_outer(outer: OuterHostInterface, path: Path) -> bool:
-    """Return True iff ``path`` is currently a mountpoint on the outer."""
-    result = outer.execute_idempotent_command(
-        f"mountpoint -q {shlex.quote(str(path))}",
-        timeout_seconds=10.0,
-    )
-    return result.success
-
-
-def _is_fstab_entry_present_on_outer(outer: OuterHostInterface, loop_file_path: Path) -> bool:
-    """Return True iff ``/etc/fstab`` already references ``loop_file_path`` at column 1."""
-    pattern = f"^{re.escape(str(loop_file_path))}[[:space:]]"
-    result = outer.execute_idempotent_command(
-        f"grep -qE {shlex.quote(pattern)} /etc/fstab",
-        timeout_seconds=10.0,
-    )
-    return result.success
-
-
-def _exec_in_container(
-    outer: OuterHostInterface,
-    container_name: str,
-    command: str,
-    timeout_seconds: float = 300.0,
-) -> str:
-    """Execute a shell command inside a running container on outer. Returns stdout.
-
-    Forces ``--workdir /`` so the exec succeeds regardless of whether the
-    image's declared ``WORKDIR`` exists at exec time. Mngr's first
-    container exec is its own sshd setup -- which runs *before* any
-    ``post_host_create_command`` hook -- so a WORKDIR like ``/mngr/code/``
-    that the image expects to be populated by a first-boot seed step
-    won't be on disk yet. None of mngr's automated setup commands depend
-    on the image's WORKDIR; they all use absolute paths.
-
-    Raises MngrError if the command exits non-zero.
-    """
-    remote = f"docker exec --workdir / {shlex.quote(container_name)} sh -c {shlex.quote(command)}"
-    result = outer.execute_idempotent_command(remote, timeout_seconds=timeout_seconds)
-    if not result.success:
-        raise MngrError(f"docker exec in {container_name} failed: {result.stderr.strip() or result.stdout.strip()}")
-    return result.stdout
-
-
-def _run_docker(
-    outer: OuterHostInterface,
-    docker_args: Sequence[str],
-    timeout_seconds: float = 60.0,
-) -> str:
-    """Run a docker subcommand on outer and return stdout.
-
-    Raises MngrError if the command exits non-zero.
-    """
-    remote = "docker " + " ".join(shlex.quote(a) for a in docker_args)
-    result = outer.execute_idempotent_command(remote, timeout_seconds=timeout_seconds)
-    if not result.success:
-        raise MngrError(f"docker {' '.join(docker_args[:2])} failed: {result.stderr.strip() or result.stdout.strip()}")
-    return result.stdout
-
-
-def _commit_container(outer: OuterHostInterface, container_name: str, image_tag: str) -> str:
-    """Commit a container to an image. Returns the image ID."""
-    return _run_docker(outer, ["commit", container_name, image_tag]).strip()
-
-
-def _stop_container(outer: OuterHostInterface, container_name: str, timeout_seconds: int = 10) -> None:
-    """Stop a running container."""
-    _run_docker(outer, ["stop", "-t", str(timeout_seconds), container_name])
-
-
-def _start_container(outer: OuterHostInterface, container_name: str) -> None:
-    """Start a stopped container."""
-    _run_docker(outer, ["start", container_name])
-
-
-def _remove_container(outer: OuterHostInterface, container_name: str, force: bool = False) -> None:
-    """Remove a container. If force=True, kill running containers first."""
-    args: list[str] = ["rm"]
-    if force:
-        args.append("-f")
-    args.append(container_name)
-    _run_docker(outer, args)
-
-
-def _remove_volume(outer: OuterHostInterface, volume_name: str) -> None:
-    """Remove a Docker named volume (force)."""
-    _run_docker(outer, ["volume", "rm", "-f", volume_name])
-
-
-def _load_resource_text(resource_name: str) -> str:
-    """Read a bundled package resource as text (e.g. snapshot_helper.sh)."""
-    return importlib_resources.files("imbue.mngr_vps_docker.resources").joinpath(resource_name).read_text()
-
-
-def _provision_snapshot_helper_on_outer(
-    outer: OuterHostInterface,
-    cg: ConcurrencyGroup,
-    *,
-    host_id: HostId,
-    btrfs_mount_path: Path,
-    subvolume_path: Path,
-    trigger_volume_name: str,
-) -> None:
-    """Install the snapshot helper systemd unit + the trigger docker volume on the outer.
-
-    Two-phase pipeline (parallelized within each phase via ``cg`` so
-    independent SSH operations don't serialize, since each one round-trips
-    over the WAN):
-
-    1. Phase A -- 5 parallel ops: write the 3 helper files + ``mkdir -p``
-       the trigger dir + ``mkdir -p`` the snapshots dir.
-    2. Phase B -- 2 parallel ops: ``systemctl daemon-reload && systemctl
-       enable --now snapshot_helper.service`` (chained into one SSH RTT)
-       and ``docker volume create`` for the trigger volume (lazy bind, so
-       safe to run in parallel with the systemctl step now that the trigger
-       dir exists).
-
-    Idempotent: rewriting the script / unit / env file is harmless; the
-    ``systemctl enable --now`` re-runs are no-ops when the unit is already
-    enabled and active; the ``docker volume create`` is no-op-with-warning
-    when the volume already exists.
-
-    Assumes ``inotify-tools`` and ``jq`` are already installed (cloud-init
-    installs them; OVH installs them via ``_REQUIRED_OUTER_PACKAGES``).
-    """
-    helper_script = _load_resource_text("snapshot_helper.sh")
-    helper_service = _load_resource_text("snapshot_helper.service")
-    helper_env = (
-        f"MNGR_BTRFS_MOUNT_PATH={btrfs_mount_path}\n"
-        f"MNGR_HOST_SUBVOLUME={subvolume_path}\n"
-        f"MNGR_TRIGGER_DIR={OUTER_SNAPSHOT_TRIGGER_DIR}\n"
-    )
-    snapshots_dir = btrfs_mount_path / "snapshots"
-
-    with log_span("Provisioning snapshot helper on outer (host_id={})", host_id):
-        with ConcurrencyGroupExecutor(
-            parent_cg=cg,
-            name="snapshot_helper_phase_a",
-            max_workers=5,
-        ) as phase_a:
-            phase_a_futures = [
-                phase_a.submit(
-                    _write_outer_file,
-                    outer,
-                    path=OUTER_HELPER_SCRIPT_PATH,
-                    content=helper_script,
-                    mode="0755",
-                ),
-                phase_a.submit(
-                    _write_outer_file,
-                    outer,
-                    path=OUTER_HELPER_SERVICE_PATH,
-                    content=helper_service,
-                    mode="0644",
-                ),
-                phase_a.submit(
-                    _write_outer_file,
-                    outer,
-                    path=OUTER_HELPER_ENV_PATH,
-                    content=helper_env,
-                    mode="0644",
-                ),
-                phase_a.submit(_ensure_outer_dir, outer, OUTER_SNAPSHOT_TRIGGER_DIR),
-                phase_a.submit(_ensure_outer_dir, outer, snapshots_dir),
-            ]
-        for future in phase_a_futures:
-            future.result()
-
-        with ConcurrencyGroupExecutor(
-            parent_cg=cg,
-            name="snapshot_helper_phase_b",
-            max_workers=2,
-        ) as phase_b:
-            phase_b_futures = [
-                phase_b.submit(_enable_snapshot_helper_unit, outer),
-                phase_b.submit(
-                    _create_bind_volume_on_outer,
-                    outer,
-                    volume_name=trigger_volume_name,
-                    device_path=OUTER_SNAPSHOT_TRIGGER_DIR,
-                ),
-            ]
-        for future in phase_b_futures:
-            future.result()
-
-
-def _enable_snapshot_helper_unit(outer: OuterHostInterface) -> None:
-    """`systemctl daemon-reload && systemctl enable --now snapshot_helper.service`.
-
-    Chained into a single SSH round-trip so this step costs one RTT
-    instead of two. The combined command is idempotent: daemon-reload
-    re-reads unit files (no effect if nothing changed) and ``enable --now``
-    is a no-op when the unit is already enabled + active.
-    """
-    result = outer.execute_idempotent_command(
-        f"systemctl daemon-reload && systemctl enable --now {OUTER_HELPER_SERVICE_NAME}",
-        timeout_seconds=20.0,
-    )
-    if not result.success:
-        raise VpsProvisioningError(
-            f"systemctl daemon-reload && systemctl enable --now "
-            f"{OUTER_HELPER_SERVICE_NAME} failed: stderr={result.stderr.strip()!r}"
-        )
-
-
-def _write_outer_file(
-    outer: OuterHostInterface,
-    *,
-    path: Path,
-    content: str,
-    mode: str,
-) -> None:
-    """Atomically write `content` to `path` on the outer with the given chmod."""
-    _ensure_outer_dir(outer, path.parent)
-    outer.write_file(path, content.encode("utf-8"), mode=mode, is_atomic=True)
-
-
-def _ensure_outer_dir(outer: OuterHostInterface, path: Path) -> None:
-    """`mkdir -p` on the outer; raises VpsProvisioningError on failure."""
-    result = outer.execute_idempotent_command(f"mkdir -p {shlex.quote(str(path))}", timeout_seconds=10.0)
-    if not result.success:
-        raise VpsProvisioningError(f"Failed to create outer dir {path}: stderr={result.stderr.strip()!r}")
-
-
-def _create_bind_volume_on_outer(
-    outer: OuterHostInterface,
-    *,
-    volume_name: str,
-    device_path: Path,
-) -> None:
-    """Create a docker named volume that bind-mounts ``device_path`` on container start.
-
-    Uses the ``local`` driver's bind options
-    (``type=none``, ``device=<path>``, ``o=bind``) so the docker volume is just
-    a name plus a record of the host path it should bind into containers; the
-    actual data lives at ``device_path`` (a btrfs subvolume here).
-    """
-    _run_docker(
-        outer,
-        [
-            "volume",
-            "create",
-            "--driver",
-            "local",
-            "--opt",
-            "type=none",
-            "--opt",
-            f"device={device_path}",
-            "--opt",
-            "o=bind",
-            volume_name,
-        ],
-    )
-
-
-def prepare_btrfs_on_outer(
-    outer: OuterHostInterface,
-    *,
-    host_id: HostId,
-    btrfs_mount_path: Path,
-    loop_file_path: Path,
-    outer_disk_reserved_gb: int,
-) -> Path:
-    """Ensure btrfs loop FS + per-host subvolume exist on the outer; return the subvolume path.
-
-    Each step is independently idempotent so a partially-failed earlier
-    ``mngr create`` (e.g. allocate succeeded, mount failed) can be retried
-    cleanly: btrfs-progs install, loop-file allocation + ``mkfs.btrfs``,
-    loop mount, ``/etc/fstab`` line append, and ``btrfs subvolume create``
-    each gate on a probe and skip when already in place. Never
-    ``mkfs.btrfs -f`` -- a populated image file is always preserved.
-
-    Returns the absolute path of the per-host subvolume on the outer,
-    suitable for use as the ``device=`` value of a bind-options docker volume.
-
-    Raises ``VpsProvisioningError`` if free space on ``/`` (after subtracting
-    ``outer_disk_reserved_gb``) is not positive, or if any setup step fails.
-    """
-    subvolume_path = btrfs_mount_path / host_id.get_uuid().hex
-
-    with log_span("Ensuring btrfs-progs is installed on outer"):
-        if not _is_btrfs_progs_installed_on_outer(outer):
-            _install_btrfs_progs_on_outer(outer)
-
-    # Allocate the loop file (if missing) sized to free-space-minus-reserve.
-    # The free-space check is skipped when the loop file already exists, so
-    # re-running on an already-provisioned VPS doesn't fail when the
-    # reserve has since been consumed by docker image layers.
-    if not _check_file_exists_on_outer(outer, loop_file_path):
-        with log_span("Computing btrfs loop file size from free space on /"):
-            free_gb = _get_outer_free_disk_gb(outer, Path("/"))
-            loop_file_size_gb = free_gb - outer_disk_reserved_gb
-            if loop_file_size_gb <= 0:
-                raise VpsProvisioningError(
-                    f"Insufficient free space on outer for btrfs loop file: "
-                    f"free={free_gb}GB, outer_disk_reserved_gb={outer_disk_reserved_gb}GB. "
-                    f"Need free > reserved. Lower outer_disk_reserved_gb or use a larger VPS plan."
-                )
-        with log_span("Allocating btrfs loop file at {} ({}GB)", loop_file_path, loop_file_size_gb):
-            mkdir_result = outer.execute_idempotent_command(
-                f"mkdir -p {shlex.quote(str(loop_file_path.parent))}",
-                timeout_seconds=10.0,
-            )
-            if not mkdir_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to create parent dir {loop_file_path.parent} for btrfs loop file: "
-                    f"stderr={mkdir_result.stderr.strip()!r}"
-                )
-            allocate_result = outer.execute_idempotent_command(
-                f"fallocate -l {loop_file_size_gb}G {shlex.quote(str(loop_file_path))}",
-                timeout_seconds=120.0,
-            )
-            if not allocate_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to fallocate btrfs loop file at {loop_file_path}: "
-                    f"stderr={allocate_result.stderr.strip()!r}"
-                )
-            mkfs_result = outer.execute_idempotent_command(
-                f"mkfs.btrfs {shlex.quote(str(loop_file_path))}",
-                timeout_seconds=180.0,
-            )
-            if not mkfs_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to mkfs.btrfs on {loop_file_path}: stderr={mkfs_result.stderr.strip()!r}"
-                )
-
-    with log_span("Mounting btrfs loop file at {}", btrfs_mount_path):
-        mkdir_mount_result = outer.execute_idempotent_command(
-            f"mkdir -p {shlex.quote(str(btrfs_mount_path))}",
-            timeout_seconds=10.0,
-        )
-        if not mkdir_mount_result.success:
-            raise VpsProvisioningError(
-                f"Failed to create btrfs mount path {btrfs_mount_path}: stderr={mkdir_mount_result.stderr.strip()!r}"
-            )
-        if not _is_path_mounted_on_outer(outer, btrfs_mount_path):
-            mount_result = outer.execute_idempotent_command(
-                f"mount -o loop {shlex.quote(str(loop_file_path))} {shlex.quote(str(btrfs_mount_path))}",
-                timeout_seconds=30.0,
-            )
-            if not mount_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to loop-mount {loop_file_path} at {btrfs_mount_path}: "
-                    f"stderr={mount_result.stderr.strip()!r}"
-                )
-
-    if not _is_fstab_entry_present_on_outer(outer, loop_file_path):
-        with log_span("Appending fstab entry for {}", loop_file_path):
-            fstab_line = f"{loop_file_path}  {btrfs_mount_path}  btrfs  loop,defaults  0 0"
-            # echo + >> is the simplest idempotency-safe form here once the grep
-            # above has confirmed the line is absent (no risk of duplicating).
-            append_result = outer.execute_idempotent_command(
-                f"echo {shlex.quote(fstab_line)} >> /etc/fstab",
-                timeout_seconds=10.0,
-            )
-            if not append_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to append fstab entry for {loop_file_path}: stderr={append_result.stderr.strip()!r}"
-                )
-
-    if not _check_directory_exists_on_outer(outer, subvolume_path):
-        with log_span("Creating btrfs subvolume {}", subvolume_path):
-            create_subvol_result = outer.execute_idempotent_command(
-                f"btrfs subvolume create {shlex.quote(str(subvolume_path))}",
-                timeout_seconds=30.0,
-            )
-            if not create_subvol_result.success:
-                raise VpsProvisioningError(
-                    f"Failed to create btrfs subvolume at {subvolume_path}: "
-                    f"stderr={create_subvol_result.stderr.strip()!r}"
-                )
-
-    return subvolume_path
-
-
-def _delete_btrfs_subvolume_on_outer(outer: OuterHostInterface, subvolume_path: Path) -> None:
-    """Delete a btrfs subvolume on the outer; raise MngrError on failure.
-
-    No-op when the subvolume's path is already absent on the outer (caller has
-    nothing else to do for an already-cleaned-up host).
-    """
-    if not _check_directory_exists_on_outer(outer, subvolume_path):
-        return
-    result = outer.execute_idempotent_command(
-        f"btrfs subvolume delete {shlex.quote(str(subvolume_path))}",
-        timeout_seconds=60.0,
-    )
-    if not result.success:
-        raise MngrError(f"btrfs subvolume delete {subvolume_path} failed: stderr={result.stderr.strip()!r}")
-
-
-def _seed_host_volume_layout_on_outer(outer: OuterHostInterface, subvolume_path: Path) -> None:
-    """Pre-create the ``host_dir/`` and ``agents/`` subdirectories of the subvolume.
-
-    A single ``mkdir -p`` so downstream writers (the agent container,
-    ``persist_agent_data``) don't need to mkdir first. Idempotent.
-    """
-    host_dir_path = subvolume_path / HOST_DIR_SUBPATH
-    agents_dir_path = subvolume_path / AGENTS_SUBPATH
-    result = outer.execute_idempotent_command(
-        f"mkdir -p {shlex.quote(str(host_dir_path))} {shlex.quote(str(agents_dir_path))}",
-        timeout_seconds=10.0,
-    )
-    if not result.success:
-        raise MngrError(f"Failed to seed host volume layout under {subvolume_path}: stderr={result.stderr.strip()!r}")
-
-
-def _pull_image(outer: OuterHostInterface, image: str, timeout_seconds: float = 300.0) -> None:
-    """Pull a Docker image."""
-    _run_docker(outer, ["pull", image], timeout_seconds=timeout_seconds)
-
-
-def _run_container(
-    outer: OuterHostInterface,
-    *,
-    image: str,
-    name: str,
-    port_mappings: Mapping[str, str],
-    volumes: Sequence[str],
-    labels: Mapping[str, str],
-    extra_args: Sequence[str],
-    entrypoint_cmd: str,
-) -> str:
-    """Run a detached docker container on outer. Returns the container id."""
-    args: list[str] = ["run", "-d", "--name", name]
-    for host_bind, container_port in port_mappings.items():
-        args.extend(["-p", f"{host_bind}:{container_port}"])
-    for vol in volumes:
-        args.extend(["-v", vol])
-    for key, value in labels.items():
-        args.extend(["--label", f"{key}={value}"])
-    args.extend(extra_args)
-    args.extend(["--entrypoint", "sh", image, "-c", entrypoint_cmd])
-    output = _run_docker(outer, args, timeout_seconds=120.0)
-    container_id = output.strip()
-    logger.debug("Started container {} ({})", name, container_id[:12])
-    return container_id
-
-
-def _build_ssh_transport_for_outer(outer: OuterHostInterface) -> tuple[str, str, str, int, str]:
-    """Build the rsync ssh-transport command and key fields for the given outer.
-
-    Returns (ssh_command, ssh_user, hostname, port, ssh_key_path_str). Raises
-    MngrError if outer has no SSH connection info (i.e. is local).
-    """
-    info = outer.get_ssh_connection_info()
-    if info is None:
-        raise MngrError("Cannot upload directory to a local outer host")
-    user, hostname, port, key_path = info
-    # Mirror docker_over_ssh._SSH_BASE_OPTIONS plus the outer host's known_hosts
-    # so rsync's ssh subprocess uses the same trust store as the outer host.
-    host_data = outer.connector.host.data
-    known_hosts = host_data.get("ssh_known_hosts_file", "")
-    ssh_cmd = (
-        f"ssh -i {shlex.quote(str(key_path))} "
-        f"-o UserKnownHostsFile={shlex.quote(str(known_hosts))} "
-        f"-o StrictHostKeyChecking=yes "
-        f"-o BatchMode=yes "
-        f"-o ConnectTimeout=15 "
-        f"-o ServerAliveInterval=20 "
-        f"-o ServerAliveCountMax=10"
-    )
-    return ssh_cmd, user, hostname, port, str(key_path)
-
-
-def _upload_directory_to_outer(
-    outer: OuterHostInterface,
-    cg: ConcurrencyGroup,
-    local_path: Path,
-    remote_path: str,
-    timeout_seconds: float = 900.0,
-) -> None:
-    """Upload a local directory to outer via rsync over SSH.
-
-    Mirrors the behavior of the legacy ``DockerOverSsh.upload_directory``:
-    retries connection-class failures (broken pipe, RST, ssh-disconnect) up
-    to ``_UPLOAD_MAX_ATTEMPTS`` with backoff, since fresh Vultr VPSes
-    routinely drop the first SSH connection in their first minute of life.
-    ``--partial-dir`` lets retries resume rather than re-upload from
-    scratch; that path lives outside the build context so partial files
-    never end up baked into the docker image. Non-retryable rsync errors
-    fail fast on the first attempt.
-    """
-    ssh_cmd, user, hostname, _port, _key_path = _build_ssh_transport_for_outer(outer)
-    local_str = str(local_path).rstrip("/") + "/"
-    cmd = [
-        "rsync",
-        "-az",
-        "--delete",
-        f"--partial-dir={_RSYNC_PARTIAL_DIR_REMOTE}",
-        "--exclude=__pycache__",
-        "--exclude=.venv",
-        "--exclude=node_modules",
-        "--exclude=.mypy_cache",
-        "--exclude=.ruff_cache",
-        "--exclude=.pytest_cache",
-        "--exclude=.test_output",
-        "--exclude=htmlcov",
-        "--exclude=.test_durations",
-        "-e",
-        ssh_cmd,
-        local_str,
-        f"{user}@{hostname}:{remote_path}/",
-    ]
-    logger.debug("Uploading {} to {}@{}:{}", local_path, user, hostname, remote_path)
-
-    last_stderr = ""
-    for attempt in range(1, _UPLOAD_MAX_ATTEMPTS + 1):
-        finished = cg.run_process_to_completion(
-            command=cmd,
-            is_checked_after=False,
-            timeout=timeout_seconds,
-        )
-        if finished.is_timed_out:
-            # Whole-process timeout: don't retry (the next attempt would
-            # just hit the same timeout again, and we'd take 3x longer
-            # to surface a real "VPS is wedged" diagnosis).
-            raise MngrError(f"Upload timed out after {timeout_seconds}s")
-        if finished.returncode == 0:
-            return
-        last_stderr = finished.stderr.strip()
-        is_last_attempt = attempt == _UPLOAD_MAX_ATTEMPTS
-        if is_last_attempt or not _is_retryable_rsync_error(last_stderr):
-            break
-        backoff_seconds = _UPLOAD_RETRY_BACKOFF_SECONDS[attempt - 1]
-        logger.warning(
-            "Upload to {} attempt {}/{} failed; retrying in {:.0f}s. stderr={!r}",
-            hostname,
-            attempt,
-            _UPLOAD_MAX_ATTEMPTS,
-            backoff_seconds,
-            last_stderr,
-        )
-        time.sleep(backoff_seconds)
-    raise MngrError(f"Upload failed: {last_stderr}")
-
-
-def _noop_line_sink(_line: str) -> None:
-    """No-op line sink for ``execute_streaming_command`` callers that don't care about output."""
-
-
-def _build_image_on_outer(
-    outer: OuterHostInterface,
-    *,
-    tag: str,
-    build_context_path: str,
-    docker_build_args: Sequence[str],
-    timeout_seconds: float,
-    on_output: Callable[[str], None] | None,
-    builder: DockerBuilder,
-) -> str:
-    """Build a Docker image on outer from a remote build context. Returns the tag.
-
-    When ``builder`` is DEPOT, ensures the depot CLI is installed on outer,
-    forwards DEPOT_TOKEN (required) from the agent's environment, optionally
-    forwards DEPOT_PROJECT_ID when set, and runs ``depot build --load``.
-    """
-    if builder is DockerBuilder.DEPOT:
-        depot_token = os.environ.get("DEPOT_TOKEN", "")
-        depot_project_id = os.environ.get("DEPOT_PROJECT_ID", "")
-        if not depot_token:
-            raise MngrError(
-                "builder=DEPOT requires DEPOT_TOKEN in the agent's environment. "
-                "Set DEPOT_TOKEN (and DEPOT_PROJECT_ID if no depot.json is on the VPS), "
-                "or set builder=DOCKER."
-            )
-        args = ["build", "--load", "-t", tag] + list(docker_build_args) + [build_context_path]
-        quoted = " ".join(shlex.quote(a) for a in args)
-        env: dict[str, str] = {"DEPOT_TOKEN": depot_token}
-        if depot_project_id:
-            env["DEPOT_PROJECT_ID"] = depot_project_id
-        remote_cmd = f"{_DEPOT_INSTALL_CMD} && depot {quoted}"
-        run_env: Mapping[str, str] | None = env
-    else:
-        args = ["build", "-t", tag] + list(docker_build_args) + [build_context_path]
-        remote_cmd = "docker " + " ".join(shlex.quote(a) for a in args)
-        run_env = None
-
-    safe_remote_cmd = _redact_secret_env(remote_cmd)
-    logger.trace("docker build remote command: {}", safe_remote_cmd)
-
-    # Stream build output line-by-line so the user sees progress during long
-    # docker builds. execute_streaming_command treats the command as
-    # idempotent and retries transient SSH errors with backoff -- on retry
-    # on_output will be re-invoked with the new attempt's output (duplicates
-    # are expected and acceptable for docker build).
-    line_callback: Callable[[str], None] = on_output if on_output is not None else _noop_line_sink
-    result = outer.execute_streaming_command(
-        remote_cmd,
-        line_callback,
-        env=run_env,
-        timeout_seconds=timeout_seconds,
-    )
-    if not result.success:
-        tail = "\n".join((result.stdout + "\n" + result.stderr).splitlines()[-50:])
-        raise MngrError(f"Remote docker build failed: {tail}")
-    return tag
 
 
 class VpsDockerProvider(BaseProviderInstance):
@@ -1218,7 +407,7 @@ class VpsDockerProvider(BaseProviderInstance):
         """Callback when host data.json is updated -- sync to the unified host volume."""
         try:
             with self._make_outer_for_vps_ip(vps_ip) as outer:
-                host_store = open_host_store(outer, _host_volume_name_for(host_id))
+                host_store = open_host_store(outer, host_volume_name_for(host_id))
                 existing = host_store.read_host_record()
                 if existing is not None:
                     updated = existing.model_copy(update={"certified_host_data": certified_data})
@@ -1234,7 +423,7 @@ class VpsDockerProvider(BaseProviderInstance):
         """Wait for cloud-init to finish (Docker installed, marker file present)."""
         start = time.monotonic()
         while time.monotonic() - start < timeout_seconds:
-            if _check_file_exists_on_outer(outer, Path("/var/run/mngr-ready")):
+            if check_file_exists_on_outer(outer, Path("/var/run/mngr-ready")):
                 elapsed = time.monotonic() - start
                 if elapsed > 30.0:
                     logger.warning("Cloud-init took {:.1f}s (threshold: 30s)", elapsed)
@@ -1260,51 +449,27 @@ class VpsDockerProvider(BaseProviderInstance):
         known_hosts_entries: tuple[str, ...],
         authorized_keys_entries: tuple[str, ...],
     ) -> None:
-        """Set up SSH inside the container via docker exec."""
-        container_key_path, container_public_key = self._get_container_ssh_keypair()
+        """Set up SSH inside the container via docker exec.
+
+        Delegates to the shared ``setup_container_ssh`` helper, supplying the
+        container client/host keypairs this provider manages. The container is
+        reached via <vps_ip>:<container_ssh_port> directly; the caller is
+        responsible for adding the matching known_hosts entry once the VPS IP
+        is known.
+        """
+        _container_key_path, container_public_key = self._get_container_ssh_keypair()
         container_host_key_path, container_host_public_key = self._get_container_host_keypair()
-        container_host_private_key = container_host_key_path.read_text()
-
-        # Install packages and set up host_dir
-        with log_span("Installing packages in container"):
-            install_cmd = build_check_and_install_packages_command(
-                mngr_host_dir=str(self.host_dir),
-                host_volume_mount_path=host_volume_mount_path,
-            )
-            _exec_in_container(outer, container_name, install_cmd, timeout_seconds=300.0)
-
-        # Configure SSH keys
-        with log_span("Configuring SSH in container"):
-            ssh_cmd = build_configure_ssh_command(
-                user="root",
-                client_public_key=container_public_key,
-                host_private_key=container_host_private_key,
-                host_public_key=container_host_public_key,
-            )
-            _exec_in_container(outer, container_name, ssh_cmd)
-
-        # Add known_hosts entries
-        known_hosts_cmd = build_add_known_hosts_command("root", known_hosts_entries)
-        if known_hosts_cmd is not None:
-            _exec_in_container(outer, container_name, known_hosts_cmd)
-
-        # Add authorized_keys entries
-        auth_keys_cmd = build_add_authorized_keys_command("root", authorized_keys_entries)
-        if auth_keys_cmd is not None:
-            _exec_in_container(outer, container_name, auth_keys_cmd)
-
-        # Start sshd
-        with log_span("Starting sshd in container"):
-            _exec_in_container(
-                outer,
-                container_name,
-                "/usr/sbin/sshd -D -o MaxSessions=100 &",
-            )
-
-        # Add container host key to local known_hosts.
-        # The container is reached via <vps_ip>:<container_ssh_port> directly.
-        # We need to add the key for that endpoint. Since we don't know the
-        # VPS IP here, the caller is responsible for adding the known_hosts entry.
+        setup_container_ssh(
+            outer,
+            container_name,
+            mngr_host_dir=str(self.host_dir),
+            host_volume_mount_path=host_volume_mount_path,
+            container_public_key=container_public_key,
+            container_host_private_key=container_host_key_path.read_text(),
+            container_host_public_key=container_host_public_key,
+            known_hosts_entries=known_hosts_entries,
+            authorized_keys_entries=authorized_keys_entries,
+        )
 
     def _prepare_btrfs_on_outer(self, outer: OuterHostInterface, host_id: HostId) -> Path:
         """Ensure btrfs loop FS + per-host subvolume exist on the outer; return the subvolume path.
@@ -1520,13 +685,13 @@ class VpsDockerProvider(BaseProviderInstance):
         ``_finalize_host_creation`` writes ``host_state.json`` -- even if a
         later step in this method fails.
         """
-        volume_name = _host_volume_name_for(host_id)
-        snapshot_trigger_volume_name = _snapshot_trigger_volume_name_for(host_id)
+        volume_name = host_volume_name_for(host_id)
+        snapshot_trigger_volume_name = snapshot_trigger_volume_name_for(host_id)
 
         with log_span("Provisioning unified host volume on btrfs subvolume"):
             subvolume_path = self._prepare_btrfs_on_outer(outer, host_id)
-            _seed_host_volume_layout_on_outer(outer, subvolume_path)
-            _create_bind_volume_on_outer(outer, volume_name=volume_name, device_path=subvolume_path)
+            seed_host_volume_layout_on_outer(outer, subvolume_path)
+            create_bind_volume_on_outer(outer, volume_name=volume_name, device_path=subvolume_path)
 
         # Snapshot helper: lets the in-container host_backup service request
         # `btrfs subvolume snapshot` against the per-host subvolume via a
@@ -1534,10 +699,10 @@ class VpsDockerProvider(BaseProviderInstance):
         # volume. Both the systemd unit on the outer and the docker volume
         # mounted at /mngr-snapshot/ in the container need to exist before
         # the agent boots. The helper provisioning is internally a 2-phase
-        # parallel pipeline (see _provision_snapshot_helper_on_outer) so
+        # parallel pipeline (see provision_snapshot_helper_on_outer) so
         # the ~7 SSH round-trips it would otherwise serialize collapse to
         # the latency of 2.
-        _provision_snapshot_helper_on_outer(
+        provision_snapshot_helper_on_outer(
             outer,
             self.mngr_ctx.concurrency_group,
             host_id=host_id,
@@ -1551,7 +716,7 @@ class VpsDockerProvider(BaseProviderInstance):
         else:
             logger.log(LogLevel.BUILD.value, "Pulling Docker image {} on VPS...", base_image, source="vps")
             with log_span("Pulling Docker image on VPS"):
-                _pull_image(outer, base_image, timeout_seconds=300.0)
+                pull_image(outer, base_image, timeout_seconds=300.0)
 
         container_name = f"{self.mngr_ctx.config.prefix}{name}"
         labels = {
@@ -1563,7 +728,7 @@ class VpsDockerProvider(BaseProviderInstance):
         logger.log(LogLevel.BUILD.value, "Starting Docker container on VPS...", source="vps")
         snapshots_dir_on_outer = self.config.btrfs_mount_path / "snapshots"
         with log_span("Starting Docker container"):
-            container_id = _run_container(
+            container_id = run_container(
                 outer,
                 image=base_image,
                 name=container_name,
@@ -1655,7 +820,7 @@ class VpsDockerProvider(BaseProviderInstance):
         self._create_shutdown_script(host)
         with log_span("Starting activity watcher"):
             start_watcher_cmd = build_start_activity_watcher_command(str(self.host_dir))
-            _exec_in_container(outer, container_name, start_watcher_cmd)
+            exec_in_container(outer, container_name, start_watcher_cmd)
 
         host_record = VpsDockerHostRecord(
             certified_host_data=host_data,
@@ -1721,139 +886,18 @@ class VpsDockerProvider(BaseProviderInstance):
     ) -> str:
         """Build a Docker image on the VPS from the provided build args.
 
-        Uploads the build context (if a local path is referenced) to the VPS
-        and runs docker build there. Returns the image tag to use.
-
-        If the local build context is a git worktree, clones it into a temp
-        directory first so the .git directory is self-contained. If git_depth
-        is specified, the clone uses --depth and always creates a temp clone
-        (even for non-worktree repos).
+        Thin wrapper around the shared ``build_image_on_outer_from_build_args``
+        helper, supplying this provider's configured docker builder. Returns the
+        image tag to use.
         """
-        build_tag = f"mngr-build-{host_id}"
-        remote_build_dir = f"/tmp/mngr-build-{host_id.get_uuid().hex}"
-
-        # Separate the build context path from other docker build args.
-        # Docker build expects the last positional arg to be the context path.
-        # We scan for args that look like local paths (not starting with --)
-        # and upload them as the build context.
-        context_args: list[str] = []
-        non_context_args: list[str] = []
-        for arg in docker_build_args:
-            if not arg.startswith("-") and Path(arg).exists():
-                context_args.append(arg)
-            else:
-                non_context_args.append(arg)
-
-        # If the build context is a git worktree or --git-depth is set,
-        # clone into a temp directory to get a standalone .git directory.
-        local_clone_dir: Path | None = None
-        if context_args:
-            local_context = Path(context_args[-1]).resolve()
-            is_worktree = (local_context / ".git").is_file()
-            if is_worktree or git_depth is not None:
-                local_clone_dir = Path(tempfile.mkdtemp(prefix="mngr-vps-build-"))
-                clone_reason = "worktree" if is_worktree else f"--git-depth={git_depth}"
-                logger.log(
-                    LogLevel.BUILD.value,
-                    "Cloning build context locally ({})...",
-                    clone_reason,
-                    source="vps",
-                )
-                clone_cmd = ["git", "clone"]
-                if git_depth is not None:
-                    clone_cmd.extend(["--depth", str(git_depth)])
-                # Use file:// so --depth is honored for local repos
-                clone_target = local_clone_dir / "repo"
-                clone_cmd.extend([f"file://{local_context}", str(clone_target)])
-                cg = ConcurrencyGroup(name="git-clone-build-context")
-                with cg:
-                    clone_result = cg.run_process_to_completion(
-                        command=clone_cmd,
-                        is_checked_after=False,
-                        timeout=120.0,
-                    )
-                    if clone_result.returncode != 0:
-                        raise MngrError(f"Failed to clone build context: {clone_result.stderr.strip()}")
-                    # Overlay the worktree's working tree on top of the
-                    # fresh clone so uncommitted edits (e.g. a locally-
-                    # rsynced ``vendor/mngr/`` from
-                    # ``mngr imbue_cloud admin pool create --mngr-source``,
-                    # or any in-flight FCT edits the operator hasn't
-                    # committed yet) actually reach the docker build.
-                    # ``git clone`` alone only carries committed files,
-                    # which silently rolls the build context back to
-                    # HEAD and produces "Unknown field" / "wrong code"
-                    # surprises at image-build time. Mirrors the
-                    # corresponding overlay step in minds' agent_creator
-                    # (Apr 12 fix, commit a3dc008b4); the shared helper
-                    # in ``imbue.mngr.utils.git_utils`` keeps the two
-                    # call sites in lockstep.
-                    if is_worktree:
-                        rsync_worktree_over_clone(local_context, clone_target, cg=cg)
-                context_args[-1] = str(clone_target)
-
-        try:
-            logger.log(
-                LogLevel.BUILD.value,
-                "Building Docker image on VPS (this may take several minutes)...",
-                source="docker",
-            )
-            if context_args:
-                upload_context = Path(context_args[-1])
-                logger.log(LogLevel.BUILD.value, "Uploading build context to VPS...", source="vps")
-                with log_span("Uploading build context to VPS"):
-                    mkdir_result = outer.execute_idempotent_command(f"mkdir -p {shlex.quote(remote_build_dir)}")
-                    if not mkdir_result.success:
-                        raise MngrError(
-                            f"Failed to create remote build dir {remote_build_dir}: {mkdir_result.stderr.strip()}"
-                        )
-                    upload_cg = ConcurrencyGroup(name="rsync-build-context")
-                    with upload_cg:
-                        _upload_directory_to_outer(outer, upload_cg, upload_context, remote_build_dir)
-
-                # Rewrite --file/--dockerfile paths to absolute paths on the VPS.
-                # These are relative to the local build context, but on the VPS
-                # the context lives at remote_build_dir.
-                resolved_build_args = _resolve_dockerfile_paths(non_context_args, remote_build_dir)
-
-                with log_span("Building Docker image on VPS"):
-                    _build_image_on_outer(
-                        outer,
-                        tag=build_tag,
-                        build_context_path=remote_build_dir,
-                        docker_build_args=tuple(resolved_build_args),
-                        timeout_seconds=600.0,
-                        on_output=_emit_docker_build_output,
-                        builder=self.config.builder,
-                    )
-            else:
-                # No local context -- pass all args to docker build with a minimal context
-                mkdir_result = outer.execute_idempotent_command(f"mkdir -p {shlex.quote(remote_build_dir)}")
-                if not mkdir_result.success:
-                    raise MngrError(
-                        f"Failed to create remote build dir {remote_build_dir}: {mkdir_result.stderr.strip()}"
-                    )
-                with log_span("Building Docker image on VPS"):
-                    _build_image_on_outer(
-                        outer,
-                        tag=build_tag,
-                        build_context_path=remote_build_dir,
-                        docker_build_args=tuple(docker_build_args),
-                        timeout_seconds=600.0,
-                        on_output=_emit_docker_build_output,
-                        builder=self.config.builder,
-                    )
-            logger.log(LogLevel.BUILD.value, "Docker image built successfully", source="docker")
-        finally:
-            if local_clone_dir is not None:
-                shutil.rmtree(local_clone_dir, ignore_errors=True)
-
-        # Clean up remote build directory
-        cleanup_result = outer.execute_idempotent_command(f"rm -rf {shlex.quote(remote_build_dir)}")
-        if not cleanup_result.success:
-            logger.debug("Failed to clean up remote build dir: {}", cleanup_result.stderr.strip())
-
-        return build_tag
+        return build_image_on_outer_from_build_args(
+            outer,
+            self.mngr_ctx.concurrency_group,
+            host_id=host_id,
+            docker_build_args=docker_build_args,
+            git_depth=git_depth,
+            builder=self.config.builder,
+        )
 
     def _create_shutdown_script(self, host: Host) -> None:
         """Create the shutdown script that stops the container on idle."""
@@ -1901,7 +945,7 @@ class VpsDockerProvider(BaseProviderInstance):
 
         with self._make_outer_for_vps_ip(host_record.vps_ip) as outer:
             with log_span("Stopping container on VPS"):
-                _stop_container(outer, host_record.config.container_name, timeout_seconds=int(timeout_seconds))
+                stop_container(outer, host_record.config.container_name, timeout_seconds=int(timeout_seconds))
 
             # Update host record
             host_store = open_host_store(outer, host_record.config.volume_name)
@@ -1928,7 +972,7 @@ class VpsDockerProvider(BaseProviderInstance):
 
         with self._make_outer_for_vps_ip(host_record.vps_ip) as outer:
             with log_span("Starting container on VPS"):
-                _start_container(outer, host_record.config.container_name)
+                start_container(outer, host_record.config.container_name)
 
         # Wait for sshd in container
         with log_span("Waiting for container SSH"):
@@ -1965,7 +1009,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 # Stop and remove the agent container; removing the volume below
                 # will fail otherwise because the container still holds it open.
                 try:
-                    _remove_container(outer, vps_config.container_name, force=True)
+                    remove_container(outer, vps_config.container_name, force=True)
                 except MngrError as e:
                     logger.warning("Failed to remove container: {}", e)
 
@@ -1976,7 +1020,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 # re-runs `mngr destroy` after VPS termination has failed).
                 subvolume_path = self.config.btrfs_mount_path / host_id.get_uuid().hex
                 try:
-                    _delete_btrfs_subvolume_on_outer(outer, subvolume_path)
+                    delete_btrfs_subvolume_on_outer(outer, subvolume_path)
                 except MngrError as e:
                     logger.warning("Failed to delete btrfs subvolume {}: {}", subvolume_path, e)
 
@@ -1985,7 +1029,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 # entry still needs cleanup so a later create with the same
                 # volume name doesn't collide.
                 try:
-                    _remove_volume(outer, vps_config.volume_name)
+                    remove_volume(outer, vps_config.volume_name)
                 except MngrError as e:
                     logger.warning("Failed to remove host volume: {}", e)
 
@@ -1993,7 +1037,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 # the bind source at OUTER_SNAPSHOT_TRIGGER_DIR is shared across
                 # all containers on this outer and is left alone).
                 try:
-                    _remove_volume(outer, _snapshot_trigger_volume_name_for(host_id))
+                    remove_volume(outer, snapshot_trigger_volume_name_for(host_id))
                 except MngrError as e:
                     logger.warning("Failed to remove snapshot trigger volume: {}", e)
 
@@ -2014,11 +1058,11 @@ class VpsDockerProvider(BaseProviderInstance):
         # Clean up local known_hosts
         if vps_ip is not None:
             try:
-                _remove_host_from_known_hosts(self._vps_known_hosts_path(), vps_ip, 22)
+                remove_host_from_known_hosts(self._vps_known_hosts_path(), vps_ip, 22)
             except Exception as e:
                 logger.trace("Failed to clean up VPS known_hosts: {}", e)
             try:
-                _remove_host_from_known_hosts(
+                remove_host_from_known_hosts(
                     self._container_known_hosts_path(), vps_ip, self.config.container_ssh_port
                 )
             except Exception as e:
@@ -2075,7 +1119,7 @@ class VpsDockerProvider(BaseProviderInstance):
         if vps_ip is not None and host_record.config is not None:
             with self._make_outer_for_vps_ip(vps_ip) as outer:
                 # Check if container is running
-                if _docker_inspect_running(outer, host_record.config.container_name):
+                if docker_inspect_running(outer, host_record.config.container_name):
                     return self._create_host_object(
                         host_id, HostName(host_record.certified_host_data.host_name), vps_ip
                     )
@@ -2116,7 +1160,7 @@ class VpsDockerProvider(BaseProviderInstance):
             # Cache the host object
             if record.vps_ip is not None and record.config is not None:
                 with self._make_outer_for_vps_ip(record.vps_ip) as outer:
-                    if _docker_inspect_running(outer, record.config.container_name):
+                    if docker_inspect_running(outer, record.config.container_name):
                         self._create_host_object(host_id, host_name, record.vps_ip)
                     else:
                         self._create_offline_host(record)
@@ -2160,7 +1204,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 if container_name not in self._container_running_cache:
                     try:
                         with self._make_outer_for_vps_ip(record.vps_ip) as outer:
-                            self._container_running_cache[container_name] = _docker_inspect_running(
+                            self._container_running_cache[container_name] = docker_inspect_running(
                                 outer, container_name
                             )
                     except MngrError as exc:
@@ -2256,7 +1300,7 @@ class VpsDockerProvider(BaseProviderInstance):
                 if host_id is None:
                     logger.debug("No mngr container on VPS {} yet, skipping", vps_ip)
                     return [], {}
-                host_store = open_host_store(outer, _host_volume_name_for(host_id))
+                host_store = open_host_store(outer, host_volume_name_for(host_id))
                 record = host_store.read_host_record()
                 if record is None:
                     logger.debug("No host record on VPS {} volume yet, skipping", vps_ip)
@@ -2387,7 +1431,7 @@ class VpsDockerProvider(BaseProviderInstance):
 
             with self._make_outer_for_vps_ip(host_record.vps_ip) as outer:
                 with log_span("Collecting listing data via single SSH command"):
-                    raw_output = _exec_in_container(
+                    raw_output = exec_in_container(
                         outer,
                         host_record.config.container_name,
                         script,
@@ -2617,7 +1661,7 @@ class VpsDockerProvider(BaseProviderInstance):
 
         with self._make_outer_for_vps_ip(host_record.vps_ip) as outer:
             with log_span("Creating Docker snapshot"):
-                image_id = _commit_container(outer, host_record.config.container_name, image_tag)
+                image_id = commit_container(outer, host_record.config.container_name, image_tag)
 
             # Store snapshot record in host data
             snapshot_record = SnapshotRecord(
@@ -2665,7 +1709,7 @@ class VpsDockerProvider(BaseProviderInstance):
 
         with self._make_outer_for_vps_ip(host_record.vps_ip) as outer:
             try:
-                _run_docker(outer, ["rmi", str(snapshot_id)])
+                run_docker(outer, ["rmi", str(snapshot_id)])
             except MngrError as e:
                 logger.warning("Failed to delete snapshot image: {}", e)
